Move education Address component to common
