+++ src/js/edu-benefits/1990/components/personal-information/SecondaryContactFields.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
+import Address from '../../../../common/components/questions/Address';
 import ErrorableTextInput from '../../../../common/components/form-elements/ErrorableTextInput';
 import ErrorableCheckbox from '../../../../common/components/form-elements/ErrorableCheckbox';
 import Phone from '../../../../common/components/questions/Phone';
-import Address from '../Address';
 
 export default class SecondaryContactFields extends React.Component {
   render() {
